feat : fix argument resolver

+++ src/main/java/se/ton/t210/dto/ExpectScoreResponse.java
@@ -6,10 +6,10 @@ import lombok.Getter;
 public class ExpectScoreResponse {
 
     private final int currentScore;
-    private final int currentPercentile;
-    private final int expectedPassPercent;
+    private final float currentPercentile;
+    private final float expectedPassPercent;
 
-    public ExpectScoreResponse(int currentScore, int currentPercentile, int expectedPassPercent) {
+    public ExpectScoreResponse(int currentScore, float currentPercentile, float expectedPassPercent) {
         this.currentScore = currentScore;
         this.currentPercentile = currentPercentile;
         this.expectedPassPercent = expectedPassPercent;
